Simplify to_s
We should raise early, to avoid holding on to instances that are in an
invalid state.  The `initialize` method raises errors as soon as we know
we have bad information.

We also simplify the namespace when we include the newly created module.

We match the style of asking if something is valid by using
`valid_operators?` similar to how we decided to ask about valid
operands.