✨ iterations 9 and 10
**dchuquilla**
I think I finally get the way to use each_with_object

Now the approach is to cross horizontal and vertical sequences, to get the points that I need.

___
**kotp**

The str argument for the block is not strength, and while it is a string, it is a line, right? So we can use line to communicate this, otherwise we have to imagine what that "str" is. Not having to translate from "programming language" to the "language of the problem being solved" helps us with our cognitive load as a human.

Using @index directly is likely something we do not want to do, and is @index "line number" when we think about it? This will, if we are thinking we should be using attributes, document the code better as well.

And yes, each_with_object is is being used well.

We can also return an enum instead of to_a and that will help mean that we can do with_object.

So that would mean we can do vertical_sequence.with_object('').

Vertical sequence is probably lines, and horizontal sequence is columns, I think. Does that read better? And @index is probably line_number.

Looking forward to the next iteration!

___
**dchuquilla**
Thanks for the names of sequences that have more sense now.

I used depth to replace index, according to diamond parts definitions.

👥 Reviewers:
- KOTP