✨ Diamond iteration 11 kotp
Looking good.

We can now use columns.with_object instead of columns.each_with_object, because Enumerator instances have a with_object method that can be used.

We should probably have an instance variable for @columns, not @_columns, and the same for lines. Not sure why the name shift happened with the underscore. Can you explain? (I am sure it happened for a reason, I just do not know what that reason is.)

We are at the point where we probably want to organize our code better.

Remember that private does not affect the visibility of constants, only things at that same level of invocation of that method. So instance methods, but not constants, and not class level methods.

And having the public depth and to_s is not necessary, we only need the public to_s method.

But not only that, we currently define that method in two places. Remove the attr_reader :depth and then we will have only the one definition.

We should use the methods we created, instead of using @depth, we have a method for that…

I am looking forward to the next iteration!

dchuquilla

wow, with_object blows my mind.

I use @_columns ||= as a memoize technique, it is used only ins the method and I do not want to re-calculate in all loop iterations. I tried runing the test 5k times for 'Z' letter, it is 1 sec slower without this.

This applies also for @depth

👥 Reviewers:
- KOTP