✨ 14th iteration
Can these memoized methods be done in initialize where we are meaning to initialize our internal state? Then the attribute_reader line can manage this method creation itself?

And I think that will mean that we do not need the instance variable @letter at all.

I had a hard time finding the public method to_s as it was above the private method initialize and that was below a class level item, the constant.

Seems that the code is in a bit of disorganized state in terms of scope and even visibility.

The private does not make the constant private, so we still have to do the statement three lines after. So moving that private down should be useful, bringing it closer to the things that it effects.